requirements.txt
pvlib>=0.10.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
openpyxl>=3.1.0
scipy>=1.11.0
streamlit>=1.32.0
plotly>=5.18.0
